requirements.txt
streamlit==1.25.0
tensorflow==2.12.0
numpy==1.24.3
pillow==9.5.0
pillow-heif==0.12.0

